### 2021-10-21 #3840 @whedon
[PRE REVIEW]: MetaWards: A flexible metapopulation framework for modelling disease spread
**Submitting author:** @chryswoods (<a href="http://orcid.org/0000-0001-6563-9903">Christopher Woods</a>)
**Repository:** <a href="https://github.com/metawards/MetaWards" target ="_blank">https://github.com/metawards/MetaWards</a>
**Version:** v1.6.1
