### 2021-10-21 #3840 @chryswoods
Thanks for considering our submission. I've looked at the list of possible reviewers and maybe acolum, klmedeiros, barryrowlingson, stensy or ibontama? (I am looking at the intersection of epidemiology and python)

Thanks again for your time. Please let me know if there is anything I can do to help.
