### 2021-11-12 #3840 @majensen
@amoeba - would you have time/interest in reviewing [this work](#3840) for JOSS? Much appreciated as ever -
