### 2021-11-12 #3840 @majensen
@acolum - Would you have time and interest in reviewing [this work](#3840) for [JOSS](https://joss.theoj.org)? Really appreciate your help in the past. Thanks -@majensen
