### 2022-02-06 #3914 @majensen
@chryswoods This is a really nice paper. "Tutorial-driven development" is a great idea, your team ought to write that up sometime.
To publish, I'll need you to please create an open archive of the code repo. You can do this with Zenodo, FigShare or other service. [Here](https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729) is a little tutorial on pushing to Zenodo.
JOSS prefers that the title and author list on this archive match that of the paper.

Please let me know the DOI of the resulting archive here in this thread.
Thanks very much!
