### 2022-02-08 #3914 @chryswoods
Thanks - we've thought about writing it up, but are unsure where would be a good place to get it published.

We have the connection set up to Zenodo so that an archive and DOI are created every time we make a release. The DOI for 1.6.1 is [10.5281/zenodo.5562738](https://doi.org/10.5281/zenodo.5562738), while the "general" DOI for all releases is [10.5281/zenodo.5562738](https://doi.org/10.5281/zenodo.5562737).

Do you want us to create and archive and DOI for the paper branch of the code?

The title on the DOI is just "MetaWards", without the description of what it is (as we have after the colon in the paper title). I can change that if you think that is a good idea?

I've also just noticed that we missed off Matt Keeling from the list of authors on Zenodo. I will add Matt in.
