### 2022-02-10 #3914 @majensen
@chryswoods I think this should work.
I will ask @openjournals/joss-eics to please review https://github.com/openjournals/joss-reviews/issues/3914#issuecomment-1032431728 - I would request that we use the existing MetaWards DOI as the doi of record for JOSS, even though it's title doesn't precisely match the paper's.
