### 2022-02-11 #3914 @kthyng
Hi @chryswoods! Just a few things to wrap up:

1. doi and version look good ✅
2. references don't have correct capitalization. For example, Danon et al 2021 has the country names uncapitalized. You can preserve capitalization in your .bib file with {} around characters. Please check the rest of your references too.
