### 2022-02-14 #3914 @chryswoods
Thanks @kthyng - very useful to learn that. I've updated the references and I think all of the capitalisation in the titles is ok. I've pushed the change to the joss branch in the MetaWards repo.

I've also been through the proof and it looks great :-)
